third_party/WebKit/LayoutTests/imported/wpt/web-animations/interfaces/KeyframeEffect/setTarget-expected.txt

This is a testharness.js-based test.
FAIL Test setting target before constructing the associated animation Animation is not defined
FAIL Test setting target from null to a valid target Animation is not defined
FAIL Test setting target from a valid target to null assert_equals: Value after clearing the target expected "10px" but got "50px"
FAIL Test setting target from a valid target to another target assert_equals: Value of 1st element (currently not targeted) after changing the effect target expected "10px" but got "50px"
Harness: the test ran to completion.
